+++ packages/react-querybuilder/src/utils/parseSpEL/parseSpEL.test.ts
@@ -115,7 +115,6 @@ it('handles "like" comparisons', () => {
   );
 });
 
-// TODO:
 it('negates "matches" comparisons appropriately', () => {
   testParseSpEL(
     '!(f1 matches "Test")',